Update README.md
Yay, first commit! :D This is the beginning of what should be a fairly interesting project. I mainly wanted to start this project to see if questions could be adequetely answered using nothing but string parsing / language interpretation of web pages. The hope is that eventually an AI would be able to use this to find the answers to any non-quantitative question it may have.

We shall see where the road ahead leads.

Enjoy! :)