Update queria.py
Seems to be working fairly well. It is 2:46AM and I need to wake up around 10:00 or so... so I guess it is time for bed. Next, we'll add the ability to input your own question and answers and have it populate the lists accordingly. Until then, sleep sounds pretty nice.

Goodnight.