Update answers.txt, Update queria.py, Update questions.txt, Create mult.py, Create newmult.py, Create wiki.py
I haven't touched this project in a while, so you will have to excuse the fact that I have no idea why I created mult or newmult, or what normal.py is.
Anyways, I have recently created wiki.py, which aims to better answer free-response questions in the chance that someone doesn't have a quizlet with the appropriate answer (or in this case, the existence of the actual question is more important...)
Because the functions that make it work reside in a different file, and because I felt queria.py was getting a little too disorganized, I also reworked the code of the welcome screen. No, the welcome screen still looks the same, I just moved each of the "multiple choice" and "free response" options to their own little function. In the future I imagine each path will take on their file, or at least the free response section, as I feel that will be the most difficult one to do.

TODO: I need to figure out what the "mult" files do. I believe they had something to do with importing questions, which reminds me that I need to finish that as well (or at least perfect it).
TODO: I want to possibly add a webgui to view the answers to my questions. At the very least, have it construct an html file for me to view the answers in, as it is really difficult to read off of this terminal sometimes.
TODO: Experiment with googling questions EXACTLY as they are written (with quotation marks) and prioritizing those pages in searches.